recommendationservice/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

numpy<2.0.0 
pandas==2.2.2
scikit-learn==1.4.2
scikit-surprise==1.1.3

boto3==1.34.82

confluent-kafka==2.3.0

psycopg2-binary==2.9.9

python-dotenv==1.0.1

pydantic==2.6.4
